+++ test/integration/recommendPage.js
@@ -44,7 +44,7 @@ describe('Recommend page', () => {
     iExpect.htmlWith200Status(res);
     const $ = cheerio.load(res.text);
 
-    expect($($('.page-section p')[2]).text()).to.equal('This is because you have symptoms.');
+    expect($($('.page-section p')[3]).text()).to.equal('This is because you have symptoms.');
   });
 
   it('page title should be \'We recommend that you\' if age question is answered 15 or younger', async () => {
@@ -66,7 +66,7 @@ describe('Recommend page', () => {
     iExpect.htmlWith200Status(res);
     const $ = cheerio.load(res.text);
 
-    expect($($('.page-section p')[2]).text()).to.equal('This is because you’re under the legal age of consent for sex.');
+    expect($($('.page-section p')[3]).text()).to.equal('This is because you’re under the legal age of consent for sex.');
   });
 
   it('the page should have a link to the Location search page with the right params for symptoms', async () => {