+++ CHANGELOG.md
@@ -4,6 +4,9 @@
 - ES integration
 - Error messages tweaking
 - Search result pages design
+- Remove in app breadcrumbs
+- Skip to page content links & accessibility help pages
+- Accessibility improvements
 
 0.3.0 / 2018-03-19
 =======