+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Remove in app breadcrumbs
 - Skip to page content links & accessibility help pages
 - Accessibility improvements
+- Results opening times table
 
 0.3.0 / 2018-03-19
 =======